Update pgsql.py
Removed function of creating database on a boolean test of it existing or not, this was causing an error at run time. I will be checking to see a work around, so that new users who might not know how to work with PGSQL will not have to really worry about dealing with the db creation.